Match 68 battle-menu, action-sequence, NPC, editor, and texture-animation functions (#10)
* Match texture animation group initialization

* Match texture animation player construction

* Match texture animation group reset

* Match texture animation group enable

* Match resetting all texture animation groups

* Document texture animation group layout and remaining dependencies

* Define action sequence records and queue state

* Implement CActionSeq::ClearSeq

* Implement CActionSeq::GetNextSeq

* Implement CActionSeq::Initialize

* Implement CActionSeq constructor

* Implement CActionSeq::NextMoveSeq

* Implement CActionSeq::NextMotionSeq

* Implement CActionSeq::NextAnimeSeq

* Name action sequence operation tags

* Implement CActionSeq::MoveSeq

* Implement CActionSeq::MoveSeq

* Implement CActionSeq::SetPos

* Implement CActionSeq::RotRefSeq

* Implement CActionSeq::RotAngleSeq

* Implement CActionSeq::RotMoveSeq

* Implement CActionSeq::ClearRotSeq

* Implement CActionSeq::WaitRotSeq

* Implement CActionSeq::SetRot

* Implement CActionSeq::SetDelayRot

* Implement CActionSeq::MotionTrg

* Implement CActionSeq::AnimeTrg

* Implement CActionSeq::MotionSeq

* Implement CActionSeq::AnimeSeq

* Implement CActionSeq::MoveSeq

* Implement CActionSeq::CheckEnd

* Implement CActionSeq::SyncChara

* Implement action sequence record release

* Document action sequence parameters and playback behavior

* Match editarea part attributes and map river endpoints

* Match editable map road connection encoding

* Match editable map river connection encoding

* Define integer grid vectors and match world-coordinate altitude

* Match editarea GetAlt_i coordinate wrapper

* Match editarea SearchPartsID coordinate wrapper

* Match editarea SearchPartsExtra coordinate wrapper

* Match editarea GetGrid coordinate wrapper

* Match editable map initialization

* Match unique map-part collection in grid rectangles

* Match polygon gathering around world coordinates

* Match polygon gathering inside a world-space box

* Match editable map part bounds with retail extrema behavior

* Name proven map connection attributes and shapes

* Move matched editarea function reference assembly

* Document editarea matching results and remaining dependencies

* Define NPC movement sequence records and playback state

* Implement CNPCharacter::ClearSeq

* Implement CNPCharacter::GetNowSeq

* Implement CNPCharacter::CheckSeq

* Implement CNPCharacter::NextSeq

* Implement CNPCharacter::SetSeq

* Implement CNPCharacter::SetWait

* Define the NPC state byte cleared by initialization

* Implement CNPCharacter::ShadowStep

* Implement CNPCharacter::DrawShadow

* Implement CNPCharacter::CheckDraw

* Implement CNPCharacter::PickUpPoly

* Implement CNPCharacter::Initialize

* Implement CNPCharacter::Draw

* Implement CNPCharacter::Step

* Implement CCharacter::PickUpPoly

* Implement CNPCharacter::PlaySeq

* Implement CNPCharacter::GetNextSeq

* Document NPC playback and collision query behavior

* Clarify grid snapping coordinate documentation

* Define battle menu weapon option message state

Describe the option flags, weapon and message pointers used by MenuClsMes, and declare the shared East King message window in its owning header.

* Decompile MenuClsMes::InitData

Reset the weapon option display coordinates, mode, flags and message counts.

* Decompile MenuClsMes::Step

Refresh weapon option messages and advance the window in its active display modes.

* Decompile DrawBtlAtoraSelect

Draw the Atora selection screen through its owning menu renderer and keep the battle wrapper file-local.

* Decompile BtlDrawOption

Draw the option screen and disable bilinear filtering afterwards.

* Decompile WeaponMenuKastumSelectUp

Find the preceding enabled customization row while preserving the top-row wrap.

* Decompile WeaponMenuKastumSelectDown

Find the following enabled customization row and wrap to the top when no later row is enabled.

* Decompile GetMenuIconPos

Position each battle menu icon on the two sides of the menu ring.

* Correct WhoIsWeaponEquip return contract

Return the weapon definition owner as an integer, matching battle menu callers that consume the full result without byte sign extension.

* Decompile EnableWeaponElemNone

Identify the weapons that restrict selecting no element, including Ruby weapons and the four special guns.

* Document remaining weapon message matching blockers

Record the candidate, relocation dependency and comparison spellings, and omit the unused East King global declaration.

* Format battle menu additions and record sweep coverage

Align the new message-state fields and function boundaries, and distinguish compiler-judged drafts from twelve jump-table decompiler failures.

* Match weapon option message state updates

* Reconcile weapon option notes with the matched implementation

---------

+++ include/menu_inventory.hpp
@@ -261,13 +261,13 @@ void SeitonAttachBoard(ATTACH_LIST *attachments);
 int WhatIsKindofItem(int item_no);
 
 /**
- * Returns the character currently equipping a weapon identifier.
+ * Returns the character associated with a weapon definition, or -1 if unavailable.
  *
  * @mangled WhoIsWeaponEquip__Fi
  * @address 0x00231E90
  * @size 0x74
  */
-s8 WhoIsWeaponEquip(int weapon_no);
+int WhoIsWeaponEquip(int weapon_no);
 
 /**
  * Returns the number of attachment holes on a weapon definition.
